+++ .storybook/preview.tsx
@@ -1,0 +1,24 @@
+import type { Preview, Decorator, ReactRenderer } from '@storybook/react';
+import React from 'react';
+import { MemoryRouter } from 'react-router-dom';
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <MemoryRouter initialEntries={['/']}>
+        <Story />
+      </MemoryRouter>
+    ),
+  ],
+};
+
+export default preview;
